refactor: move custom filter logic to QueryPost class
